Rename 'metric_helpers.py' to 'stats.py'
